moved signal restore and heuristic log sanatize
