+++ lib/ansible/module_utils/basic.py
@@ -149,7 +149,7 @@ from ansible.module_utils.common._collections_compat import (
     Set, MutableSet,
 )
 from ansible.module_utils.common.locale import get_best_parsable_locale
-from ansible.module_utils.common.process import get_bin_path
+from ansible.module_utils.common.process import get_bin_path, restore_signal_handlers
 from ansible.module_utils.common.file import (
     _PERM_BITS as PERM_BITS,
     _EXEC_PERM_BITS as EXEC_PERM_BITS,
@@ -308,71 +308,6 @@ def get_all_subclasses(cls):
 
 
 # End compat shims
-
-
-def heuristic_log_sanitize(data, no_log_values=None):
-    ''' Remove strings that look like passwords from log messages '''
-    # Currently filters:
-    # user:pass@foo/whatever and http://username:pass@wherever/foo
-    # This code has false positives and consumes parts of logs that are
-    # not passwds
-
-    # begin: start of a passwd containing string
-    # end: end of a passwd containing string
-    # sep: char between user and passwd
-    # prev_begin: where in the overall string to start a search for
-    #   a passwd
-    # sep_search_end: where in the string to end a search for the sep
-    data = to_native(data)
-
-    output = []
-    begin = len(data)
-    prev_begin = begin
-    sep = 1
-    while sep:
-        # Find the potential end of a passwd
-        try:
-            end = data.rindex('@', 0, begin)
-        except ValueError:
-            # No passwd in the rest of the data
-            output.insert(0, data[0:begin])
-            break
-
-        # Search for the beginning of a passwd
-        sep = None
-        sep_search_end = end
-        while not sep:
-            # URL-style username+password
-            try:
-                begin = data.rindex('://', 0, sep_search_end)
-            except ValueError:
-                # No url style in the data, check for ssh style in the
-                # rest of the string
-                begin = 0
-            # Search for separator
-            try:
-                sep = data.index(':', begin + 3, end)
-            except ValueError:
-                # No separator; choices:
-                if begin == 0:
-                    # Searched the whole string so there's no password
-                    # here.  Return the remaining data
-                    output.insert(0, data[0:begin])
-                    break
-                # Search for a different beginning of the password field.
-                sep_search_end = begin
-                continue
-        if sep:
-            # Password was found; remove it.
-            output.insert(0, data[end:prev_begin])
-            output.insert(0, '********')
-            output.insert(0, data[begin:sep + 1])
-            prev_begin = begin
-
-    output = ''.join(output)
-    if no_log_values:
-        output = remove_values(output, no_log_values)
-    return output
 
 
 def _load_params():
@@ -1779,9 +1714,7 @@ class AnsibleModule(object):
         return self._clean
 
     def _restore_signal_handlers(self):
-        # Reset SIGPIPE to SIG_DFL, otherwise in Python2.7 it gets ignored in subprocesses.
-        if PY2 and sys.platform != 'win32':
-            signal.signal(signal.SIGPIPE, signal.SIG_DFL)
+        Restore_signal_handlers()
 
     def run_command(self, args, check_rc=False, close_fds=True, executable=None, data=None, binary_data=False, path_prefix=None, cwd=None,
                     use_unsafe_shell=False, prompt_regex=None, environ_update=None, umask=None, encoding='utf-8', errors='surrogate_or_strict',
@@ -1931,7 +1864,7 @@ class AnsibleModule(object):
             st_in = subprocess.PIPE
 
         def preexec():
-            self._restore_signal_handlers()
+            restore_signal_handlers()
             if umask:
                 os.umask(umask)
